+++ frontend/main/src/RarlaTodo/components/TodoInput/index.scss
@@ -1,0 +1,14 @@
+.todo-input {
+  box-sizing: border-box;
+  width: 600px;
+  height: 50px;
+  margin: 50px;
+  padding: 0 10px;
+  font-size: 20px;
+  text-align: center;
+  font-weight: bold;
+  color: blueviolet;
+  border: none;
+  border-bottom: 4px solid blueviolet;
+  outline: none;
+}
