+++ frontend/main/src/RarlaTodo/index.scss
@@ -4,20 +4,7 @@
   align-items: center;
   flex-direction: column;
   
-  .todo-input {
-    box-sizing: border-box;
-    width: 600px;
-    height: 50px;
-    margin: 50px;
-    padding: 0 10px;
-    font-size: 20px;
-    text-align: center;
-    font-weight: bold;
-    color: blueviolet;
-    border: none;
-    border-bottom: 4px solid blueviolet;
-    outline: none;
-  }
+  
   
   .todo-box {
     box-sizing: border-box;
